pyproject.toml
[tool.poetry]
name = "ocr"
version = "0.1.0"
description = ""
authors = ["Eryk666 <olejarzeryk@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^2.2.4"
pillow = "^11.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
